requirements_dev.txt
pip>=23.0
bumpversion>=0.6.0
wheel>=0.41.0
flake8>=6.0
tox>=4.0
coverage>=7.0
Sphinx>=7.0
pytest>=7.0
coveralls>=3.3
pytest-cov>=4.0
GitPython>=3.1.32
